+++ proofs/lang/extraction.v
@@ -25,7 +25,7 @@ Extract Constant expr.var_info => "Location.t".
 Extract Constant expr.InstrInfo.t => "IInfo.t".
 Extract Constant expr.InstrInfo.witness => "IInfo.dummy".
 Extract Constant expr.instr_info => "IInfo.t".
-Extract Constant expr.fun_info => "BinNums.positive".
+Extract Constant expr.fun_info => "FInfo.t".
 
 Cd  "lang/ocaml".